refactor(orders): wave-3 slice 3 valid_work_tiles scenario tables (#3949)
Migrate nine getValidWorkOrderTileKeys/suggestWorkOrders part shards to
table-driven scenarios under test/orders/support/suggestion/ with a thin
family runner, preserving coverage and description lineage.
